AI/requirements.txt
Flask
Pillow
pandas
flask-cors
keras
numpy
scikit-learn
joblib